changed int to long on all offset values

+++ tweet.c
@@ -3,7 +3,7 @@
 #include <string.h>
 #include "tweet.h"
 
-#define HEADER (sizeof(int))	// Size of the data before the first tweet
+#define HEADER (sizeof(long))	// Size of the data before the first tweet
 
 struct tweet{
 
@@ -21,22 +21,22 @@ void writeTweet(char *filename, TWEET *tweet){
 	return;
 }
 
-TWEET *readTweet(char *filename, int offset){
+TWEET *readTweet(char *filename, long offset){
 	return NULL;
 }
 
-int *findOffsetByUser(char *filename, char *username, int *foundOccurences){
+long *findOffsetByUser(char *filename, char *username, long *foundOccurences){
 	*foundOccurences = 0;
 	FILE *f = fopen(filename, "r");
 	if (f == NULL) { printf("Erro na leitura do arquivo.\n"); return NULL; }
 	
 	TWEET *tt = NULL;
-	int *listOffset;
-	int offset = HEADER;		// Offset of the last begin of tweet
-	int nextTweet;				// Offset from the last begin of tweet until the next begin of tweet
+	long *listOffset;
+	long offset = HEADER;		// Offset of the last begin of tweet
+	long nextTweet;				// Offset from the last begin of tweet until the next begin of tweet
 	
 	fseek(f, offset, SEEK_SET);
-	while(fread(&nextTweet, sizeof(int), 1, f)) {
+	while(fread(&nextTweet, sizeof(long), 1, f)) {
 		if (nextTweet > 0){
 			tt = readTweet(filename, offset);
 			if (!strcmp(username, tt->userName)) {
@@ -47,22 +47,22 @@ int *findOffsetByUser(char *filename, char *username, int *foundOccurences){
 			free(tt);
 		}
 		
-		offset += abs(nextTweet) + sizeof(int);
+		offset += abs(nextTweet) + sizeof(long);
 	}
 	
 	fclose(f);
 	return listOffset;
 }
 
-int *findOffsetByFavoriteCount(char *filename, int favoriteCount, int *foundOccurences){
+long *findOffsetByFavoriteCount(char *filename, int favoriteCount, long *foundOccurences){
 	return 0;
 }
 
-int *findOffsetByLanguage(char *filename, char* language, int *foundOccurences){
+long *findOffsetByLanguage(char *filename, char* language, long *foundOccurences){
 	return 0;
 }
 
-int removeTweet(char *filename, int offset){
+int removeTweet(char *filename, long offset){
 	return 0;
 }
